Changed Order for Switching Rounds
Changed the order of operations so that the round change makes more sense. Also, I'm going to need to make markers for the goal and launcher placements, as well as reset the ball. I think I'll use 'Alpha' to make the things invisible so that they're always hidden from rendering.