Added Results UI
The results UI has been added to the game, so now when the game is over you can choose to go back to the menu or just quit. I also used visibility instead of removing it from the parent, which is far better, really. I'm not sure if I'll bother to change the other ones.

I still need to set up the retry button and fix the launcher. I also need to make sure the UI scales properly.